requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-jazzmin==2.5.0
django-multiselectfield==0.1.12
idna==3.3
Pillow==9.2.0
psycopg2==2.9.3
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
